# Description

Please describe the change you have made.

## Checklist

- [ ] Tests added/updated.
- [ ] Run Demo Job Locally.
- [ ] Documentation updated.
- [ ] Changelogs updated in
[CHANGELOG.cdf-tk.md](https://github.com/cognitedata/toolkit/blob/main/CHANGELOG.cdf-tk.md).
- [ ] Template changelogs updated in
[CHANGELOG.templates.md](https://github.com/cognitedata/toolkit/blob/main/CHANGELOG.templates.md).
- [ ] Version bumped.

[_version.py](https://github.com/cognitedata/toolkit/blob/main/cognite/cognite_toolkit/_version.py)
and

[pyproject.toml](https://github.com/cognitedata/toolkit/blob/main/pyproject.toml)
per [semantic versioning](https://semver.org/).

+++ module_upgrade/run_check.py
@@ -1,3 +1,4 @@
+# ruff: noqa: E402
 import contextlib
 import os
 import platform
@@ -9,6 +10,9 @@ from contextlib import contextmanager
 from pathlib import Path
 from unittest.mock import patch
 
+# Hack to make the script work as running cdf modules upgrade
+sys.argv = ["cdf", "modules", "upgrade"]
+
 import yaml
 from dotenv import load_dotenv
 from packaging.version import Version
@@ -16,12 +20,11 @@ from packaging.version import parse as parse_version
 from rich import print
 from rich.panel import Panel
 
-from cognite_toolkit._cdf_tk.commands import BuildCommand, DeployCommand
+from cognite_toolkit._cdf_tk.commands import BuildCommand, DeployCommand, ModulesCommand
+from cognite_toolkit._cdf_tk.commands import _cli_commands as CLICommands
+from cognite_toolkit._cdf_tk.commands._changes import ManualChange
 from cognite_toolkit._cdf_tk.constants import ROOT_MODULES, SUPPORT_MODULE_UPGRADE_FROM_VERSION
 from cognite_toolkit._cdf_tk.loaders import LOADER_BY_FOLDER_NAME
-from cognite_toolkit._cdf_tk.prototypes.commands import ModulesCommand
-from cognite_toolkit._cdf_tk.prototypes.commands._changes import ManualChange
-from cognite_toolkit._cdf_tk.prototypes.commands.modules import CLICommands
 from cognite_toolkit._cdf_tk.utils import CDFToolConfig, module_from_path
 from cognite_toolkit._version import __version__
 
@@ -165,7 +168,8 @@ def run_modules_upgrade(
                 / "data_models"
                 / "4.Pump.view.yaml"
             )
-            pump_view.write_text(pump_view.read_text().replace("external_id", "externalId"))
+            if pump_view.exists():
+                pump_view.write_text(pump_view.read_text().replace("external_id", "externalId"))
 
         build = BuildCommand(print_warning=False)
         build.execute(False, project_path, build_path, build_env_name="dev", no_clean=False)
